refactor: put experiments in one module
