+++ tests/daaf/policyeval/test_results_agg_pipeline.py
@@ -4,7 +4,7 @@ import numpy as np
 import ray
 import ray.data
 
-from daaf.evalexps import results_agg_pipeline
+from daaf.experiments.evalexps import results_agg_pipeline
 
 
 def test_pipeline():
